Split tests in modules
