+++ mix.exs
@@ -9,6 +9,7 @@ defmodule LambdaEthereumConsensus.MixProject do
       start_permanent: Mix.env() == :prod,
       deps: deps(),
       dialyzer: dialyzer(),
+      elixirc_path: compiler_paths(Mix.env()),
       preferred_cli_env: [
         dialyzer: :test
       ]
@@ -35,9 +36,12 @@ defmodule LambdaEthereumConsensus.MixProject do
     ]
   end
 
-  defp dialyzer() do
+  defp dialyzer do
     [
       plt_file: {:no_warn, "priv/plts/project.plt"}
     ]
   end
+
+  defp compiler_paths(:test), do: ["test/spec-test/runners"] ++ compiler_paths(:prod)
+  defp compiler_paths(_), do: ["lib"]
 end
